Lecture 120

+++ backend/routes/posts.js
@@ -44,7 +44,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -73,13 +74,19 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      console.log(result);
-      res.status(200).json({
-        message: "Updated successfully!"
-      });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if(result.nModified > 0) {/* Quer dizer que um objeto foi modificado */
+        res.status(200).json({
+          message: "Updated successfully!"
+        });
+      } else {
+        res.status(401).json({
+          message: "Unauthorized!"
+        });
+      }
     });
   }
 );
@@ -119,11 +126,17 @@ router.get("", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(() => {
-    console.log("Deleted!");
-    res.status(200).json({
-      message: "Post deleted successfully!"
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(() => {
+    if(result.n > 0) {/* Quer dizer que um objeto foi DELETADO */
+      console.log("Deleted!");
+      res.status(200).json({
+        message: "Post deleted successfully!"
+      });
+    } else {
+      res.status(401).json({
+        message: "Unauthorized!"
+      });
+    }
   });
 });
